more work on config_modify
